refactoring

+++ src/main/java/sillenceSoft/schedulleCall/Service/AccessService.java
@@ -1,15 +1,12 @@
 package sillenceSoft.schedulleCall.Service;
 
-import io.jsonwebtoken.Claims;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-import sillenceSoft.schedulleCall.Dto.AccessListResponseDto;
-import sillenceSoft.schedulleCall.Dto.UserDto;
 import sillenceSoft.schedulleCall.Repository.AccessRepository;
-import sillenceSoft.schedulleCall.Repository.UserRepository;
+
 
 import java.util.List;
 
@@ -17,28 +14,35 @@ import java.util.List;
 @RequiredArgsConstructor
 public class AccessService {
     private final AccessRepository accessRepository;
-    private final UserRepository userRepository;
     private final SHA_256 sha256;
 
     public ResponseEntity canAccess (Long userNo, String accessUserPhone) {
             try {
                 String encrypt = sha256.encrypt(accessUserPhone);
-                //중복체크 해야함
-                accessRepository.saveAccess(userNo, encrypt);
-                return new ResponseEntity("success",HttpStatus.OK);
+                if (checkIfPresent(userNo, encrypt)>0) {
+                    return new ResponseEntity("이미 숨김해제가 되어 있습니다",HttpStatus.CONFLICT);
+                }
+                else {
+                    accessRepository.saveAccess(userNo, encrypt);
+                    return new ResponseEntity("success", HttpStatus.OK);
+                }
             }
             catch (Exception e) {
                 e.printStackTrace();
                 return new ResponseEntity(e.toString(),HttpStatus.INTERNAL_SERVER_ERROR);
             }
-
     }
 
     public ResponseEntity cannotAccess (Long userNo, String accessUserPhone) {
         try {
             String encrypt = sha256.encrypt(accessUserPhone);
-            accessRepository.deleteAccess(userNo, encrypt);
-            return new ResponseEntity("success", HttpStatus.OK);
+            if (checkIfPresent(userNo, encrypt)==0) {
+                return new ResponseEntity("이미 숨김이 되어 있습니다", HttpStatus.CONFLICT);
+            }
+            else {
+                accessRepository.deleteAccess(userNo, encrypt);
+                return new ResponseEntity("success", HttpStatus.OK);
+            }
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -48,14 +52,16 @@ public class AccessService {
 
     public ResponseEntity getAccessList(Long userNo) {
         try {
-
             List<String> accessList = accessRepository.getAccessList(userNo);
             return new ResponseEntity(accessList,HttpStatus.OK);
         }catch (Exception e) {
             e.printStackTrace();
             return new ResponseEntity(e.toString(),HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
 
+    public Long checkIfPresent (Long userNo, String accessUserPhone) {
+        return accessRepository.checkAccessOrNot(userNo,accessUserPhone);
     }
 
 }
